requirements.txt
RPi.GPIO
 picamera2 
customtkinter 
opencv-python 
Pillow